+++ Changelog.md
@@ -1,4 +1,6 @@
 --------------------------Changelog for v1.9.12-beta1 of SecurityCraft--------------------------
 
 - Fix: Module descriptions of security sea boats do not show up correctly
+- Fix: Holding a camera monitor makes it possible to identify disguised cameras
+- Fix: Items that can have linked positions (e.g. Camera Monitor) show the idle animation even when they have positions and are either dropped, or held by another entity
 - Misc.: Moved display case textures to the `textures/entity/display_case/` subfolder and renamed them